Allow grouping by Arel::Nodes::NamedFunction
This only fails because Arel::Nodes::NamedFunction responds to #name
(but not #relation) and calling #as modifies the node, resulting in
GROUP BY … AS …, which is invalid SQL.

This is primarily being done to benefit Squeel. It used to work as it
previously passed a Squeel::Node::Function that doesn't exhibit the
above behaviour but that had other problems.